# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ariadne Daemon API",
    "description": "REST API of ariadned, the coding-agent orchestrator daemon.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/v1/goals": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "List goals.",
        "operationId": "list",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status: one status, or several comma-separated\n(`status=active,completed`), matching goals in any of them.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "example": "active,completed"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Create a goal on registered repositories; the orchestrator session is\nspawned by the scheduler once agent execution lands.",
        "description": "The repos are referenced, not copied: whatever `POST /v1/repositories`\nvalidated about a checkout holds for every goal that names it, and an edit\nthere moves this goal too.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": "no such repository or orchestrator profile"
          }
        }
      }
    },
...
    "/v1/goals/{id}": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Inspect a goal.",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Delete a finished goal and everything under it.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "the goal is not finished yet; cancel it first"
          }
        }
      }
    },
    "/v1/goals/{id}/cancel": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Cancel a goal.",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/v1/goals/{id}/complete": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Complete the goal: it moves active -> completed and every session of it\nis torn down.",
        "description": "The orchestrator's call, because it is the only agent that knows whether\nthe plan did what the goal asked for — the daemon can see that every task\nended, and not whether the goal is met. The user's too: it is their goal,\nand a goal whose orchestrator will not start is otherwise one nothing can\nclose.\n\nWhat the daemon checks is the part it can see. A goal with a task still\ngoing is not one anybody may declare finished, however sure they are.",
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/v1/goals/{id}/finalize": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Finalize the plan: goal moves planning -> active and its tasks start. The\norchestrator's call alone, and there is nothing left for the user to\napprove.",
        "operationId": "finalize",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/v1/goals/{id}/messages": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "The messages of a goal: what its agents said that was not about one task.",
        "description": "A goal's channel is the orchestrator's inbox. Everything an agent says to\nit about a task is on that task instead, and this is where the rest goes.",
        "operationId": "list_goal_messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to_agent_id",
            "in": "query",
            "description": "Only the messages for this staffed agent.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "undelivered",
            "in": "query",
            "description": "Only the ones that have not reached their agent yet.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "deliver",
            "in": "query",
            "description": "Hand the caller its own undelivered messages: narrow the list to the\ncalling session's agent, and stamp every row it returns delivered.\n\nThis is a delivery, not a read. It needs a session behind it, and it\nis what makes a read of the channel count the same as a prompt: a\nmessage handed over here is never handed over again.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Send a message about the goal itself.",
        "operationId": "post_goal_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/repositories": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "List repositories.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "repositories"
        ],
        "summary": "Create a repository.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDto"
                }
              }
            }
          },
          "400": {
            "description": "not an absolute path, not a git work tree, an unknown branch, or a forge pin that is no catalog model"
          },
          "409": {
            "description": "this path and base branch are already registered, `ai` was asked for while the AI permission model is off, or the forge integration cannot be enabled"
          }
        }
      }
    },
...
    "/v1/tasks/{id}/step/complete": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteStepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": "step_gate_failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompleteGoalRequest": {
        "type": "object",
        "description": "Body of `POST /v1/goals/{id}/complete`: the orchestrator says the goal is\ndone. Its call, not the user's, and it carries nothing — every task being\nfinished or cancelled is the whole of the argument, and the daemon checks\nthat itself.",
        "additionalProperties": false
      },
      "CompleteStepRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "merge_commit": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CreateGoalRequest": {
        "type": "object",
        "required": [
          "title",
          "repository_ids",
          "model"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "effort": {
            "type": [
              "string",
              "null"
            ],
            "description": "The reasoning effort to run that model at, one of the efforts `GET\n/v1/models` lists for it; anything else is refused. Omitted (or\n\"default\") = whatever the agent runs the model at.",
            "example": "high"
          },
          "issue_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "landing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Landing",
                "description": "How every task of the goal ends. Omitted = `merge`. It cannot change\nonce the goal is created."
              }
            ]
          },
          "model": {
            "type": "string",
            "description": "What the orchestrator runs on, `<agent>:<model>` — the id of an agent\nin the ACP registry and, after the `:`, the model of it:\n`codex-acp:gpt-5.3-codex`, `opencode:ollama/llama3:8b`. Required —\na model is required, and no agent default stands in for one. The model\nhalf is free text, handed to that agent as typed; a string naming no\nregistry agent is refused, and so are the empty string and the word\n\"default\".",
            "example": "codex-acp:gpt-5.3-codex"
          },
          "repository_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Ids of registered repositories (`POST /v1/repositories`); at least one."
          },
          "title": {
            "type": "string"
          },
          "workflow": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "FinalizePlanRequest": {
        "type": "object",
        "description": "Body of `POST /v1/goals/{id}/finalize`: the orchestrator ends planning and\nexecution starts. The orchestrator's call, not the user's, and it carries\nnothing — the plan is the tasks it wrote.",
        "additionalProperties": false
      },
...
      "GoalDto": {
        "type": "object",
        "required": [
          "steps",
          "id",
          "title",
          "description",
          "status",
          "orchestrated",
          "model",
          "landing",
          "repos",
          "usage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "effort": {
            "type": [
              "string",
              "null"
            ],
            "description": "The reasoning effort that model is run at, pinned like `model`. None =\nwhatever the agent runs it at on its own.",
            "example": "high"
          },
          "id": {
            "type": "string"
          },
          "issue_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "landing": {
            "$ref": "#/components/schemas/Landing",
            "description": "How every task of the goal ends, chosen when the goal was created."
          },
          "model": {
            "type": "string",
            "description": "What the orchestrator or adopted author runs on, `<agent>:<model>`:\nthe registry agent and, after the `:`, the model of it.",
            "example": "claude-acp:claude-opus-5"
          },
          "orchestrated": {
            "type": "boolean",
            "description": "Whether this goal has an orchestrator for its lifetime."
          },
          "repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalRepositoryDto"
            },
            "description": "The registered repositories the goal works in, as they stand now: a\ngoal references them, so an edit to one shows up here."
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepDto"
            }
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/GoalUsageDto",
            "description": "What the agents of this goal have spent between them."
          },
          "workflow": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "description": "Body of `POST /v1/tasks/{id}/messages` and `POST /v1/goals/{id}/messages`.\n\nWho it is *from* comes from the session header rather than the body: an\nagent cannot send a message as somebody else, and a message with no session\nbehind it is the user's.",
        "required": [
          "kind",
          "to_actor",
          "body"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/MessageKind"
          },
          "to_actor": {
            "$ref": "#/components/schemas/Actor",
            "description": "Who it is for. `orchestrator` needs no agent id — a goal has one."
          },
          "to_agent_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The staffed agent it is for, as `GET /v1/tasks/{id}` lists them.\nRequired for `author` and `reviewer`, refused for the orchestrator."
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}